+++ cli/src/pcluster/templates/cw_dashboard_builder.py
@@ -9,11 +9,12 @@
 # OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 # limitations under the License.
 from collections import defaultdict, namedtuple
-from typing import Iterable
+from typing import Iterable, List
 
 from aws_cdk import aws_cloudwatch as cloudwatch
 from aws_cdk import aws_ec2 as ec2
 from aws_cdk import aws_logs as logs
+from aws_cdk.aws_cloudwatch import IAlarm
 from aws_cdk.core import Construct, Duration, Stack
 
 from pcluster.config.cluster_config import BaseClusterConfig, ExistingFileCache, SharedFsxLustre, SharedStorageType
@@ -21,6 +22,7 @@ from pcluster.constants import Feature
 from pcluster.utils import is_feature_supported
 
 MAX_WIDTH = 24
+MIN_HEIGHT = 1
 
 
 class Coord:
@@ -66,6 +68,7 @@ class CWDashboardConstruct(Construct):
         shared_storage_infos: dict,
         cw_log_group_name: str,
         cw_log_group: logs.CfnLogGroup,
+        head_node_alarms: List[IAlarm],
     ):
         super().__init__(scope, id)
         self.stack_scope = scope
@@ -75,6 +78,7 @@ class CWDashboardConstruct(Construct):
         self.shared_storage_infos = shared_storage_infos
         self.cw_log_group_name = cw_log_group_name
         self.cw_log_group = cw_log_group
+        self.head_node_alarms = head_node_alarms
 
         self.dashboard_name = self.stack_name + "-" + self._stack_region
         self.coord = Coord(x_value=0, y_value=0)
@@ -100,6 +104,11 @@ class CWDashboardConstruct(Construct):
         self.cloudwatch_dashboard = cloudwatch.Dashboard(
             self.stack_scope, "CloudwatchDashboard", dashboard_name=self.dashboard_name
         )
+
+        # Cluster Alarms
+        if self.config.monitoring.alarms.enabled:
+            self._add_text_widget("# Cluster Alarms")
+            self._add_cluster_alarms()
 
         # Create a text widget for title "Head Node EC2 metrics"
         self._add_text_widget("# Head Node EC2 Metrics")
@@ -187,6 +196,19 @@ class CWDashboardConstruct(Construct):
         text_widget.position(x=self.coord.x_value, y=self.coord.y_value)
         self.cloudwatch_dashboard.add_widgets(text_widget)
         self._update_coord_after_section(d_y=1)
+
+    def _generate_alarms_widget(self, title: str, alarms_list: List[IAlarm], **widget_kwargs):
+        """Generate an alarms widget and update the coordinates."""
+        widget = cloudwatch.AlarmStatusWidget(
+            title=title,
+            alarms=alarms_list,
+            width=MAX_WIDTH,
+            height=MIN_HEIGHT * 2,
+            **widget_kwargs,
+        )
+        widget.position(x=self.coord.x_value, y=self.coord.y_value)
+        self._update_coord(self.graph_width, self.graph_height)
+        return widget
 
     def _generate_graph_widget(self, title, metric_list, **widget_kwargs):
         """Generate a graph widget and update the coordinates."""
@@ -507,6 +529,12 @@ class CWDashboardConstruct(Construct):
         for title, metric_list in metric_graphs.items():
             widgets_list.append(self._generate_graph_widget(title, metric_list))
         return widgets_list
+
+    def _add_cluster_alarms(self):
+        widgets_list = [self._generate_alarms_widget(title="Head Node Alarms", alarms_list=self.head_node_alarms)]
+
+        self.cloudwatch_dashboard.add_widgets(*widgets_list)
+        self._update_coord_after_section(self.graph_height)
 
     def _add_head_node_instance_metrics_graphs(self):
         # Create a text widget for subtitle "Head Node Instance Metrics"