+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGELOG
 
 **CHANGES**
 - Changed cluster alarms naming convention to '[cluster-name]-[component-name]-[metric]'.
+- Add head node alarms to cluster dashboard.
 
 **BUG FIXES**
 - Fix inconsistent configuration after cluster update rollback when modifying the list of instance types declared in the Compute Resources.